test(metrics): reconcile review inventory and package size
